test: reorganizing the test code
