+++ sql-genie-test/src/test/java/io/github/genie/sql/test/DataSourceConfig.java
@@ -1,0 +1,12 @@
+package io.github.genie.sql.test;
+
+import lombok.Data;
+
+@Data
+public class DataSourceConfig {
+
+    private String url = "jdbc:mysql:///sql-dsl";
+    private String user = "root";
+    private String password = "root";
+
+}
