+++ sql-genie-test/src/test/java/io/github/genie/sql/test/EntityManagers.java
@@ -4,8 +4,13 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.Persistence;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class EntityManagers {
-    private static final EntityManagerFactory factory = Persistence.createEntityManagerFactory("org.hibernate.jpa");
+    private static final EntityManagerFactory factory = getEntityManagerFactory();
+
+
     private static final EntityManager ENTITY_MANAGER = doGetEntityManager();
 
     private EntityManagers() {
@@ -20,4 +25,14 @@ public class EntityManagers {
     }
 
 
+    private static EntityManagerFactory getEntityManagerFactory() {
+        DataSourceConfig config = new DataSourceConfig();
+
+        Map<String, String> properties = new HashMap<>();
+        properties.put("javax.persistence.jdbc.url", config.getUrl());
+        properties.put("javax.persistence.jdbc.user", config.getUser());
+        properties.put("javax.persistence.jdbc.password", config.getPassword());
+
+        return Persistence.createEntityManagerFactory("org.hibernate.jpa", properties);
+    }
 }
