+++ sql-genie-test/src/test/resources/META-INF/persistence.xml
@@ -5,13 +5,12 @@
              version="2.0">
   <persistence-unit name="org.hibernate.jpa">
     <properties>
-      <property name="jakarta.persistence.jdbc.url" value="jdbc:mysql:///sql-dsl?rewriteBatchedStatements=true"/>
-      <property name="jakarta.persistence.jdbc.driver" value="com.mysql.cj.jdbc.Driver"/>
-      <property name="jakarta.persistence.jdbc.password" value="root"/>
-      <property name="jakarta.persistence.jdbc.user" value="root"/>
+      <!--<property name="javax.persistence.jdbc.url" value="jdbc:mysql:///sql-dsl?rewriteBatchedStatements=true"/>
+      <property name="javax.persistence.jdbc.password" value="root"/>
+      <property name="javax.persistence.jdbc.user" value="root"/>-->
+      <property name="javax.persistence.jdbc.driver" value="com.mysql.cj.jdbc.Driver"/>
       <property name="hibernate.hbm2ddl.auto" value="update"/>
       <property name="hibernate.format_sql" value="true"/>
-      <!--<property name="hibernate.show_sql" value="true"/>-->
       <property name="hibernate.jdbc.batch_size" value="1000"/>
       <property name="hibernate.physical_naming_strategy"
                 value="org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy"/>
